[Style] Merge & Archive Authentication Related File * ./Sources/Model/AuthModel.swift 파일에 관련모델 통합   : AuthenticationManager - AuthDataResultModel   : SignInGoogleHelper - GoogleSignInResultModel * 사용되지 않는 인증관련 파일 아카이빙   : ./Sources/Extensions/Auth+Extension.swift   : ./Sources/Models/GoogleUser.swift
