+++ teamplan/Sources/Model/AuthenticationModel.swift
@@ -1,0 +1,29 @@
+//
+//  AuthenticationModel.swift
+//  teamplan
+//
+//  Created by 주찬혁 on 2023/07/04.
+//  Copyright © 2023 team1os. All rights reserved.
+//
+
+import Foundation
+import FirebaseAuth
+
+struct AuthDataResultModel {
+    let uid: String
+    let email: String?
+    let photoUrl: String?
+    
+    init(user: User) {
+        self.uid = user.uid
+        self.email = user.email
+        self.photoUrl = user.photoURL?.absoluteString
+    }
+}
+
+struct GoogleSignInResultModel {
+    let idToken: String
+    let accessToken: String
+    let name: String?
+    let email: String?
+}
