+++ teamplan/Sources/Services/Authentication/AuthenticationManager.swift
@@ -9,19 +9,6 @@
 import Foundation
 import FirebaseAuth
 
-
-struct AuthDataResultModel {
-    let uid: String
-    let email: String?
-    let photoUrl: String?
-    
-    init(user: User) {
-        self.uid = user.uid
-        self.email = user.email
-        self.photoUrl = user.photoURL?.absoluteString
-    }
-}
-
 enum AuthProviderOption: String {
     case email = "password"
     case google = "google.com"
@@ -61,8 +48,6 @@ final class AuthenticationManager {
     func singOut() throws {
         try Auth.auth().signOut()
     }
-    
-
 }
 
 // MARK: - SIGN IN SSO
